% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxaudit_failcache.R
\docType{class}
\name{cxaudit_failcache-class}
\alias{cxaudit_failcache-class}
\alias{cxaudit_failcache}
\title{Utility class to manage the audit trail fail cache}
\description{
The utility class represents the audit record fail cache. The fail cache
is a disk-based storage of audit records to capture records that fail to
commit to the configured audit trail storage.

Audit fail cache is enabled by the app property \code{AUDIT.FAILCACHE} equal to
\code{enable} or \code{enabled}.

The fail cache path is set using the app property \code{AUDIT.FAILCACHE.PATH},
which must exists.

The method \code{isenabled} returns \code{TRUE} if the fail cache is enabled and the
fail cache directory path exists. Otherwise, \code{FALSE}.

The method \code{isempty} returns \code{TRUE} if the fail cache is empty. A non-empty
fail cache contains one or more JSON files (\code{.json} file extension) exists
in the sub-directories of fail cache directory \code{AUDIT.FAILCACHE.PATH}. The
method does not validate that the JSON file represents a valid audit record.

The \code{cache} method stores the specified audit records \code{x} in the fail cache.
If \code{x} is a list of audit records, the record relationships are preserved.

The method \code{push} commits the audit records to the audit record store
as specified or returned by \link[cxaudit]{cxaudit_store}. If a commit of a
record is successful, the record is deleted from the fail cache. An import
event is added to each commit of audit records.
}
\section{Fields}{

\describe{
\item{\code{.attr}}{Internal storage of fail cache attributes}
}}

\section{Methods}{

\describe{
\item{\code{cache(x)}}{Cache audit records in the audit fail cache}

\item{\code{initialize()}}{Initialize audit fail cache store}

\item{\code{isempty()}}{Assert if audit fail cache store is empty}

\item{\code{isenabled()}}{Assert if audit fail cache store is enabled}

\item{\code{push(store = cxaudit::cxaudit_store())}}{Push content of audit fail cache to audit store}
}}

