% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxaudit_get.R
\name{cxaudit_get}
\alias{cxaudit_get}
\title{Get an audit record}
\usage{
cxaudit_get(x, store = cxaudit::cxaudit_store())
}
\arguments{
\item{x}{An audit record or record ID}

\item{store}{Audit trail storage}
}
\value{
A \link[cxaudit]{cxaudit_record} object
}
\description{
Retrieve an audit record by reference, either using the audit record
identifier or \link[cxaudit]{cxaudit_record}.

If no audit record exists with reference \code{x} or an error occurred when
retrieving the record, invisible \code{NULL} is returned. An invisible \code{NULL} is
also returned if auditing is disabled.
}
