% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxaudit_list.R
\name{cxaudit_list}
\alias{cxaudit_list}
\title{List audit records}
\usage{
cxaudit_list(..., store = cxaudit::cxaudit_store())
}
\arguments{
\item{...}{Filter conditions as named parameters or as a single list of named elements}

\item{store}{Audit trail storage}
}
\value{
A list of \link[cxaudit]{cxaudit_record} entries
}
\description{
Retrieve a list of audit records that satisfy the specified criteria from the
specified/configured storage \code{store} if auditing is enabled. If auditing is
disabled or an error occurs, an invisible \code{NULL} is returned.

The following filter criteria can be specified as named entries in a list or
as named arguments. Specifying filter criteria as both a list of entries and
named arguments are not supported.

\itemize{
\item \code{events} the audited event (see \link[cxaudit]{cxaudit_record} for valid values)
\item \code{object.types} the type of the audited object
\item \code{object.classes} further classifies the object \code{type}
\item \code{object.parentpaths} the parent of the object path
\item \code{object.names} the object name
\item \code{object.hashes} the object hash
\item \code{actors} are the services and/or users associated with an audited event
\item \code{envs} as the environments associated with an audited event
\item \code{from} is start of the period for selecting audit records by their corresponding date
\item \code{to} is end of the period for selecting audit records by their corresponding date
\item \code{limit} is the number of records to retrieve
\item \code{offset} is the first record number to select from
\item \code{select} is the direction from which the number of records is taken
}

The returned record selection satisfies all filter criteria analogous to the
\emph{and} operator.

The filter criteria \code{object.parentpaths} and \code{object.names} can be used to
filter on a particular object by its path, i.e. the concatenation of object
parent path and name.

The date range specified by \code{from} and \code{to} are specified as a date or date
and time using \link[base]{Sys.Date}, \link[base]{Sys.time} or
\link[base]{as.POSIXct} or \link[base]{as.POSIXlt}. The \code{to} and \code{from}
values are converted to UTC timezone truncated to number of seconds. Partial
dates are currently not supported.

If \code{select = first}, the records selected are those in order of chronological
occurrence by record date and time.

If \code{select=last} (default), records are selected in order of occurrence by
descending date and time (starting with the last record first).

The options \code{limit} and \code{offset} can be used to only return a set number of
records and/or page through the records returned. The \code{select} option is
applied before limiting and offsetting any records.

Note that if there are multiple records for the same date and time and the
selection boundary set by \code{limit}, \code{offset} and \code{select} would mean that some
records fall within the span of \code{limit} and \code{offset} and some \emph{valid}
records fall outside of that boundary, i.e. not included, then the cut-off
is arbitrary. There is no indicator or notification that valid records were
excluded. To ensure that all the records for the given period of interest
is included, increase the number of records specified by limit to ensure that
all the records for a given period is returned.

The options \code{from}, \code{to}, \code{limit}, \code{offset} and \code{select} can only be a single
value. All other options is a vector of one or more elements.

All filter criteria are disabled with the last 300 records from the
last 30 days returned by default.

The attribute \code{filter} of the returned list of records specifies the filter
criteria resulting in the selected list. See \link[base]{attr} for details on
accessing \code{filter} attribute.

If this function is used in conjunction with pagination, the filter attributes
returned with the record selection can be used to ensure that the next page by
\code{limit} and \code{offset} are not affected by new records.
}
