% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxaudit_record.R
\docType{class}
\name{cxaudit_record-class}
\alias{cxaudit_record-class}
\alias{cxaudit_record}
\title{Reference class representing an audit record}
\description{
Represents an audit record.

The class is initialized with a vector or list of named properties.
\itemize{
\item \code{event} on the object being audited
\item \code{object.type} as the type of audited object
\item \code{object.class} as the class of the audited object type
\item \code{object.path} as a path reference to the audit object
\item \code{object.hash} as a digest or hash referring to the object definition or content
\item \code{label} as a human readable label to the audited event
\item \code{actor} meaning the user or service performing or initiating the event
\item \code{env} the environment where the audited event occurred
}

The special properties \code{id} and \code{datetime} can be assigned when the the audit
record is initialized and should only be included when the audit record
represents saved audit records.

The \code{event} is one of the supported event types. The events \code{create}, \code{read},
\code{update}, \code{delete} and \code{commit} refer to the content of the audited
object. Note that the \code{commit} event may refer to an object being created or
updated.

The \code{execute} and \code{fail} events are intended to represent audited processing
events on or with objects. The \code{fail} event is not expected to used to capture
and record errors and warnings, unless they are included as the fail record
attributes.

The events \code{lock}, \code{unlock} and \code{sign} are representative of common audited
compliance states with the first two synonymous to \emph{read-only} and
\emph{write enabled}, respectively. The \code{sign} event by itself is not by
default a verifiable sign event, but can include additional attributes to
corroborate the sign event.

The events \code{connect} and \code{disconnect} are provided to capture audited connections,
such as login's, logout's and other types and styles of connections.

An audit object is referred to by a representative path \code{object.path} such that
the last level of the path is the object name and the preceding level represents
a parent hierarchy, akin to a parent directory or folder. The levels of the path
are valid references (\link{cxaudit_validreference}).

The object can be further identified by its \code{object.type} and \code{object.class}
where the type is a generic reference to the type of object and class further
classifies the type. The type and class are valid references
(\link{cxaudit_validreference}).

The \code{object.hash} is a representation of the object definition or content such
that the same object stored under different path references or across environments
represent the same object, i.e. enabling auditing and tracking objects across
environments (think of all the file transfers in today's modern processes).

The \code{object.hash}, if not specified, is the SHA-1 digest of the string
\verb{<object.type>:<object.class>:<name>} and where \emph{name} is the last level of
\code{object.path}.

The \code{getproperties} method returns a list of the record reference named
properties.

One or more additional attributes can be recorded as part of the audit record.
An attribute is a value identified by a keyword. The attribute value can also
include additional context or scope, such as \emph{new} or \emph{old}, using
a qualifier.

The \code{setattributes} methods can be used to associate attributes with the
audit record. The vector \code{x} is a named character vector where the name of
the vector element is the attribute name and the vector element value is the
attribute value. A qualifier can be included in the entry name using the convention
\verb{<name>:<qualifier>} where both the name and qualifier are valid references
(\link{cxaudit_validreference}).

The \code{getattributes} method returns a character vector of named elements
representing the record attributes. The qualifier is omitted if it is not
defined or equal to the default \code{value}.
}
\section{Fields}{

\describe{
\item{\code{.attr}}{Internal storage of record properties and attributes}
}}

\section{Methods}{

\describe{
\item{\code{getattributes()}}{Return list of record attributes}

\item{\code{getlinks()}}{Return list of link audit records}

\item{\code{getproperties()}}{Return list of record properties}

\item{\code{initialize(x)}}{Initialize record representation}

\item{\code{setattributes(x)}}{Set one or more record attributes}

\item{\code{setproperties(x)}}{Set record properties}
}}

