% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxaudit_store.R
\name{cxaudit_store}
\alias{cxaudit_store}
\title{Utility function to return an audit record storage object}
\usage{
cxaudit_store(x)
}
\arguments{
\item{x}{Database connection (optional)}
}
\value{
A simple cxaudit storage object
}
\description{
A cxaudit storage object represents storage methods to commit, retrieve and
list audit records.

Note that below configuration uses app properties and \link[cxapp]{cxapp_config}.

If the \code{AUDIT} app property is not equal to enabled, \code{NULL} is returned.

If the \code{AUDIT} app property is equal to \code{enable} or \code{enabled}, the app
property \code{AUDIT.STORE} includes the keyword \code{database} and the app property
\code{AUDIT.DB.VENDOR} is equal to \code{sqlite} or \code{postgres}, all case insensitive,
an initialized internal relational database storage object
\link[cxaudit]{.cxaudit_rdbstore} is returned.

If a connection \code{x} is specified, the storage object returned uses the connection.
If a connection is not specified, a connection is initiated from the cxaudit
configuration parameters.

A database connection can be initiated when an existing connection is not
specified.

For SQLite, use the following app properties
\itemize{
\item \code{AUDIT.STORE} equal to \code{database}
\item \code{AUDIT.DB.VENDOR} equal to \code{sqlite}
\item \code{AUDIT.DB.PATH} where the path represents the SQLite database file
}

For Postgres, use the following app properties
\itemize{
\item \code{AUDIT.STORE} equal to \code{database}
\item \code{AUDIT.DB.VENDOR} equal to \code{postgres}
\item \code{AUDIT.DB.DRIVER} as the R package database driver. Default is
\code{RPostgreSQL::PostgreSQL()}.
\item \code{AUDIT.DB.DATABASE} as database name
\item \code{AUDIT.DB.HOST} as the database host
\item \code{AUDIT.DB.PORT} as the database host port. Default is 5432.
\item \code{AUDIT.DB.USERNAME} as the database account
\item \code{AUDIT.DB.PASSWORD} as the database account password. Note that the
app configuration supports environmental variables and vaults.
}

Database connections use a pool of connections (\link[pool]{dbPool}) for
connection resilience even though most database transactions are single
thread and relies on a single database connection. The pook configuration
can be set using the app properties

\itemize{
\item \code{AUDIT.DB.POOL.MINSIZE} as pool minimum size. Default is 1.
\item \code{AUDIT.DB.POOL.MAXSIZE} as pool maximum size. Default is 25.
\item \code{AUDIT.DB.POOL.IDLETIMEOUT} as pool connection timeout. Default is
120 seconds.
}
}
