% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxaudit_validevent.R
\name{cxaudit_validevent}
\alias{cxaudit_validevent}
\title{Assert if a specified even is valid}
\usage{
cxaudit_validevent(x)
}
\arguments{
\item{x}{Event string}
}
\value{
Logical \code{TRUE} if event is valid. \code{FALSE} otherwise
}
\description{
An event is a case insensitive keyword representing the event.

See \link{.cxaudit_eventnames} for a list of valid events.
}
