% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxaudit_validreference.R
\name{cxaudit_validreference}
\alias{cxaudit_validreference}
\title{Utility function to assert if the string is a valid reference}
\usage{
cxaudit_validreference(x)
}
\arguments{
\item{x}{A vector of references}
}
\value{
Logical vector
}
\description{
Returns a logical vector of the same length as \code{x} with \code{TRUE} if the item
is a valid reference and \code{FALSE} otherwise.

A reference is a case insensitive string consisting of
\itemize{
\item letters A-Z, digits 0-9 and punctuation dash \code{-}, underscore \verb{_}
and period \code{.}
\item starts and ends with a letter or digit
\item is between 2 and 100 characters in length
}
}
