% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cxaudit_eventnames.R
\name{.cxaudit_eventnames}
\alias{.cxaudit_eventnames}
\title{Internal vector representing valid audit events}
\usage{
.cxaudit_eventnames()
}
\value{
Character vector of event names
}
\description{
The event is represented as a case insensitive keyword, most often
lower case.

Valid events are \code{create}, \code{read}, \code{write}, \code{update}, \code{delete}, \code{execute},
\code{fail}, \code{commit}, \code{lock}, \code{unlock}, \code{sign}, \code{connect}, \code{disconnect} and
\code{import}.
}
\keyword{internal}
