% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cxaudit_propertynames.R
\name{.cxaudit_propertynames}
\alias{.cxaudit_propertynames}
\title{Internal vector represents valid property names}
\usage{
.cxaudit_propertynames()
}
\value{
Character vector of property names
}
\description{
Internal vector represents valid property names
}
\keyword{internal}
