% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-cxaudit_rdbstore.R
\docType{class}
\name{.cxaudit_rdbstore-class}
\alias{.cxaudit_rdbstore-class}
\alias{.cxaudit_rdbstore}
\title{Internal utility reference class to manage audit record database storage}
\description{
Internal utility reference class to use a database to store audit records.

The \code{list} method retrieves a list of records given the conditions \code{x}. See
\link[cxaudit]{cxaudit_list} for supported options. The results are returned
as a nested list.

The \code{commit}method takes a list \code{x} of audit records and stores them in the
database. The entries of \code{x} are of type \link[cxaudit]{cxaudit_record}.
}
\section{Fields}{

\describe{
\item{\code{.attr}}{Internal attribute storage}
}}

\section{Methods}{

\describe{
\item{\code{commit(x)}}{Commit audit records to database}

\item{\code{initialize(x)}}{Initialize relation database store}
}}

