% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxaudit_commit.R
\name{cxaudit_commit}
\alias{cxaudit_commit}
\title{Utility function to commit audit records}
\usage{
cxaudit_commit(x)
}
\arguments{
\item{x}{List of audit records}
}
\value{
Invisible logical of result
}
\description{
Commits the audit records to the auditor service instance

The auditor service connection is configured using the app properties

\itemize{
\item \code{AUDITOR.URL} defines the URL for the service
\item \code{AUDITOR.TOKEN} defines the access token for the auditor service
\item \code{AUDITOR.FAILCACHE} directory path to save audit records on connection
failure
}

See \link[cxapp]{cxapp_config} for details.
}
